allow Nano to be replaced with any other editor
Previously, removal of the Nano editor in favour of other editors would
force removal of this metapackage, causing subsequent upgrade issues.

The inclusion of Nano as a dependency is only done to ensure that an
editor is installed during the build process.*  Nano, specifically, is
not used by anything but a user.

With this change, Nano is still provided by default but users are able
to remove it in favour of other editors if desired.

*  https://discourse.osmc.tv/t/rbp2-device-osmc-package-depends-on-nano-not-editor/4311/4