[package][mediacenter-addon-osmc] Walkthru: changed network connection check process
The network check begins when the Walkthru is started, when the user
comes to the connection page a progress bar is shown if the internet
check is not complete.  (If it is complete and no connection is
available, the networking page is shown.) The connection check is given
a total of 2 minutes from instantiation to complete, with waits of 10
seconds between checks.