Fix OSMC Walkthru hanging if no network detected
OSMC Waltkthru would hang when clicking 'continue' on the license accept panel if there was no network.

This was due to originally the function call missing the controlID argument.
The quick fix for this was also causing an error due to passing 2x arguments ('self' which is not needed in Python). 

The still_checking_for_network function doesn't actually use the controlID argument, so this PR reverts the function call back to original and removes the argument from the function. There are no other calls to this function so no "breakages" elsewhere.