[package][mediacenter-addon-osmc] Pi: swapped the references from out pin to in pin
Added action so when file is called as script that it reads the config, parses it, runs it back through the parser, writes to the temp location and then overwrites the original config file. This is to apply the conversion from the legacy gpio pin structure to the new one.