find the settings that exist before setting them
This also reverts the discarding of output for the attempts to set
variables, since if they fail, that can change the result code of the
script, which leaves systemd thinking that the script failed. If this
happens it is helpful to see the reason for the failure in the journal.