[package] [ftr-osmc] Fix SSID base64 encoding
base64.b64encode function takes a bytes-like object as parameter
Previously a simple string object was passed to it, which raised a TypeError.
This resulted in a faulty preseed WiFi configuration making WiFi
basically useless when booting up OSMC, because it prevented connecting
to any kind of WiFi network manually on the GUI, too.
Added the necessary bytes conversion of the parsed SSID and also joined
the double variable assignments into one, so now it encodes, decodes the
parsed SSID and finally strips the base64 encoded string in one step.
Also fixed a typo made in the preseed provisioning file name:
conmann -> connman